cast strlen as aunsigned in my if statment

+++ more_malloc_free/1-string_nconcat.c
@@ -16,7 +16,7 @@ char *string_nconcat(char *s1, char *s2, unsigned int n)
 	int i;
 	int j;
 	
-	if (n > _strlen(s2))
+	if (n > (unsigned int)_strlen(s2))
 	{
 		n = _strlen(s2);
 	}
